icache can cater to buswidths larger than its response width
the response width of the i-cache is typically 32-bits. Earlier this was also the request packet size sent on the bus even if the bus was 64-bits wide. 
This commit ensure that the cache can request burst sizes equal to the bus-width size while still catering to a 32-bit response on the core side.