Removed replay logic and added BypassFIFO to store tags from first stage
Implementing the replay logic results in an SRAM to SRAM path. Also, when a replay happens, new instruction from the core cannot be serviced in that cycle. Instead, having a BypassFIFO to store the tag would result in correct functionality, and also lets new request from core to be serviced in the cycle after fill buffer starts releasing. This would result in greater throughput of the cache.