Added store responses
In the current version, store responses are always sent in the second stage (provided the pipeline is flowing). Also, if MSHR wants to send a response to core, and in the same cycle a store response needs to be sent, the store request is stalled untill the MSHR responses are finished. 
In the later versions, once a completion store buffer (CSB) is implemented, store responses can have a different interface that directly sends the response to  CSB. Therefore, these stalls will be avoided.